Fix bug when the user tried to manage shipment exceptions.
There as a crash but after fixing the crash we raised a user error which
prevented lines to be updated if they had more than one stock move. The check
was correct but should only be applied if the user modified any fields, not
when the update is done by the exception management wizard.

Now the check takes into account which fields have been modified before trying
to raise the user error.